Simulator and Java API:
Implemented UNSOLICITEDSTATUS auth type handshake protocol semantic. The
simualtor sends an UNSOLICITEDSTATUS message to client library that
connects to it and sets "connectionID", "clusterVersion", etc in it.

Client library implementation should wait/block the connection creation
process until the UNSOLICITEDSTATUS message is received. 

Client library is required to use the connectionID in all the
subsequential request messages.  Any request message sent without the
connectionID will be rejected (not yet enfornced, but will do so when
drive has the behavior).

Simulator sets the connectionID in each of the response message.