Java API/Simulator:
A pin operation is authenticated with PINAuth authentication method.  No
user identity/Hmac information (HMACAuth) is set to the protocol
message.

A pin operation MUST be performed in a TLS connection.  Otherwise,
INVALID_REQUEST is set to the response message and the service request
is not performed in the simulator.