Calculate a serial number for each simulator instance. This makes it possible to use the serial number to identify an instance of simulator during and after restarted from a shutdown/crash.
A simulator instance serial number is calculated as follows (separated
by '-'.

SERIAL_PREFIX + ip + khomeHash + persistHome

where SERIAL_PREFIX is a constant,
ip is the ip of the simualtor,
khomeHash is the home (hash code) of the simulator instance,
persistHome is the persist home name of the instance set in the
SimulatorConfiguration.